Re #2181: Suppressed a couple of compilation warnings.
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@6008 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip-apps/src/pjsua/pjsua_app_legacy.c
@@ -649,9 +649,9 @@ static void vid_handle_menu(char *menuin)
 			   (int)info.name.slen, info.name.ptr,
 			   22-(int)info.name.slen, "                   ",
 			   s,
-			   20-pj_ansi_strlen(s), "                    ",
+			   20-(int)pj_ansi_strlen(s), "                    ",
 			   tr_list,
-			   12-pj_ansi_strlen(tr_list), "            ",
+			   12-(int)pj_ansi_strlen(tr_list), "            ",
 			   li_list);
 		}
 	    }
